// admin/types/demo.d.ts
type InventoryStatus = 'INSTOCK' | 'LOWSTOCK' | 'OUTOFSTOCK';
// ...
declare namespace Demo {
    interface Task {
        id?: number;
        name?: string;
        description?: string;
        completed?: boolean;
        status?: string;
        comments?: string;
        attachments?: string;
        members?: Member[];
        startDate?: string;
        endDate?: string;
    }

    type Category  = {
        id?:  string | int;
        name: string;
        short_name: string;
        description: string;
        handle: string;
        image: File | null;
        [key: string]: any;
    }

    interface Brand {
        name: string;
        id?: string;
        short_name?: string;
        description?: string;
        image?: string
        handle: string;
    }
// ...
    type Product = {
        id?: string;
        sku: string;
        handle?:string;
        name: string;
        description?: string;
        images?: File[] | string[];
        price: number;
        cost: number;
        stock: number;
        stock_alert?: number;
        brand_id?: string;
        brand?: Brand;
        inventoryStatus?: InventoryStatus;
        featured: boolean;
        tax_rate?: number;
        shipping_required?: boolean;
        free_shipping?: boolean;
        thumbnail?: File | null | string;
        collection_id?: string;
        is_active:  boolean;
        is_giftcard:  boolean;
        rating?: number;
        categories?: Category[];
        [key: string]: any;
    };
// ...
}